Remove game components
