--- web/src/app/game/page.module.scss
@@ -1,14 +1,0 @@
-.game_container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-
-  padding: 50px;
-
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  gap: 20px; /* Espace entre les éléments */
-}
