--- web/src/app/page.module.scss
@@ -1,6 +1,0 @@
-.intro_container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
