--- web/src/components/GameView/DisplayScreen/styles.module.scss
@@ -1,6 +1,0 @@
-.displayScreen {
-  background-color: #333;
-  color: white;
-  padding: 20px;
-  border-radius: 0px;
-}
