--- web/src/components/GameView/DoomScrollGame/styles.module.scss
@@ -1,3 +1,0 @@
-.threadmill__container {
-  flex: 1;
-}
