--- web/src/components/GameView/IntroView/styles.module.scss
@@ -1,15 +1,0 @@
-.game_intro {
-  display: flex;
-  gap: 20px;
-  flex-direction: column;
-
-  section {
-    & > * {
-      margin: 10px;
-    }
-    input {
-      display: block;
-      margin: auto;
-    }
-  }
-}
