--- web/src/components/GameView/Leaderboard/index.module.scss
@@ -1,59 +1,0 @@
-.leaderboard {
-  color: black;
-  display: flex;
-  flex-direction: column;
-}
-
-.score_board {
-  background-color: white;
-  border: 1px solid #ccc;
-  border-radius: 12px;
-  padding: 10px;
-  width: 400px;
-
-  h3 {
-    margin: 0;
-    padding: 0;
-    font-size: 18px;
-    text-align: center;
-  }
-
-  ul {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-  }
-
-  li {
-    margin: 5px 0;
-    font-size: 16px;
-  }
-}
-
-.rewards_popup {
-  background-color: white;
-  border: 1px solid #ccc;
-  margin-top: 20px;
-}
-
-.rewards-content {
-  h3 {
-    margin: 0;
-    padding: 0;
-    font-size: 18px;
-    text-align: center;
-  }
-
-  ul {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-    display: flex; /* Affiche les récompenses sur une ligne */
-    flex-wrap: wrap; /* Permet de passer à la ligne suivante si nécessaire */
-  }
-
-  li {
-    margin: 5px 10px; /* Ajustez l'espacement entre les récompenses */
-    font-size: 16px;
-  }
-}
