--- web/src/components/GameView/Menu/styles.module.scss
@@ -1,8 +1,0 @@
-.menu {
-  position: fixed;
-  bottom: 0;
-  width: 100%;
-
-  display: flex;
-  justify-content: center;
-}
