--- web/src/components/GameView/Threadmill/Carpet/styles.module.scss
@@ -1,23 +1,0 @@
-.container {
-  user-select: none;
-  pointer-events: none;
-
-  position: absolute;
-
-  left: 10%;
-  top: 5%;
-  bottom: 5%;
-  width: 80%;
-
-  overflow: hidden;
-}
-
-.scroller {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-
-  img {
-    width: 100%;
-  }
-}
