--- web/src/components/GameView/Threadmill/styles.module.scss
@@ -1,18 +1,0 @@
-.running_track {
-  width: 100%;
-  height: 100%;
-  position: relative;
-  display: inline-block;
-  overflow: hidden;
-  margin: 0;
-
-  & > img {
-    display: block;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    height: 100%;
-    min-width: 100%;
-    transform: translate(-50%, -50%);
-  }
-}
