--- web/src/components/GameView/styles.module.scss
@@ -1,10 +1,0 @@
-.game__container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-
-  display: flex;
-  flex-direction: column;
-}
