--- web/src/components/Layout/Overlay/index.tsx
@@ -1,8 +1,0 @@
-import {PropsWithChildren, FC} from 'react';
-import styles from './styles.module.scss';
-
-const Overlay: FC<PropsWithChildren> = ({children}) => {
-  return <div className={styles.overlay}>{children}</div>;
-};
-
-export default Overlay;
