--- web/src/components/Layout/Overlay/styles.module.scss
@@ -1,17 +1,0 @@
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  color: white;
-  text-align: center;
-
-  background-color: #000b;
-  backdrop-filter: blur(10px);
-}
