+++ web/src/components/Overlay/Trophy/animation.ts
