+++ web/src/components/Overlay/Trophy/index.tsx
