+++ web/src/components/Overlay/Trophy/styles.module.scss
