--- web/src/components/Three/Scenes/BabylonDoomscroll/Treadmill/Material/Shaders/fragment.ts
@@ -1,1 +1,0 @@
-export const fragmentShader = /* glsl */ ``;
