--- web/src/components/Three/Scenes/BabylonDoomscroll/Treadmill/Material/Shaders/vertex.ts
@@ -1,1 +1,0 @@
-export const vertexShader = /* glsl */ ``;
