+++ web/src/config/media.ts
