--- web/src/config/rewardsConfig.ts
@@ -1,6 +1,0 @@
-export const REWARDS = [
-  { distance: 150, name: 'Titre de techserf de bronze' },
-  { distance: 300, name: 'Titre de techserf d\'argent' },
-  { distance: 500, name: 'Titre de techserf d\'or' },
-  { distance: 1000, name: 'Titre de Techlord du Doomscroll' },
-];
